+++ st2common/st2common/models/db/__init__.py
@@ -1,3 +1,18 @@
+# Licensed to the StackStorm, Inc ('StackStorm') under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import copy
 import importlib
 
@@ -45,8 +60,8 @@ def get_model_classes():
 
 def db_setup(db_name, db_host, db_port, username=None, password=None,
              ensure_indexes=True):
-    LOG.info('Connecting to database "%s" @ "%s:%s" as user "%s".' %
-             (db_name, db_host, db_port, str(username)))
+    LOG.info('Connecting to database "%s" @ "%s:%s" as user "%s".',
+             db_name, db_host, db_port, str(username))
     connection = mongoengine.connection.connect(db_name, host=db_host,
                                                 port=db_port, tz_aware=True,
                                                 username=username, password=password)